Backtester/requirements.txt
# Core
backtrader
pandas
numpy

# API & Services
python-dotenv
fyers-apiv3
groq

# LangChain & RAG
langchain
langchain-core
langchain-groq
langchain-community
faiss-cpu
sentence-transformers

# UI
streamlit